Replace rand-elt with core rand-nth and fix bug
One of the examples at the bottom of the buffer fails in modern Clojure
versions because the `a` and `b` in this function can be the same node
and thus the literal map in the body of the `let` would have duplicate
keys, which causes an exception to be thrown.

This fix ensures the selection of `b` is drawn from all nodes with `a`
excluded.